Inherit BaseSendEmailService from waste-carriers-engine
