+++ app/services/notify/renewal_reminder_email_service.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Notify
-  class RenewalReminderEmailService < BaseSendEmailService
+  class RenewalReminderEmailService < ::WasteCarriersEngine::Notify::BaseSendEmailService
     private
 
     def notify_options
@@ -21,6 +21,16 @@ module Notify
 
     def renewal_fee
       display_pence_as_pounds(Rails.configuration.renewal_charge)
+    end
+
+    def validate_contact_email(registration)
+      raise Exceptions::MissingContactEmailError, registration.reg_identifier unless registration.contact_email.present?
+
+      assisted_digital_match = registration.contact_email == WasteCarriersEngine.configuration.assisted_digital_email
+
+      raise Exceptions::AssistedDigitalContactEmailError, registration.reg_identifier if assisted_digital_match
+
+      registration.contact_email
     end
   end
 end
